pyproject.toml
[tool.poetry]
name = "mp-blog"
version = "0.1.0"
description = "Package for munichpavel blog posts"
authors = ["Paul Larsen <munichpavel@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mp_blog", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
openai = "^0.27.8"
notebook = "^6.5.4"
pyrate-limiter = "^2.10.0"
pyarrow = "^12.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
